# Copy label values to avoid unsafe memory use