This is an alternative solution to #2000 (and #2004) where we do more copying on the error path but need fewer comments to avoid mistakes.

Benchmarks show this alternative is significantly more expensive on the error path:
Before #2000 
```
BenchmarkIngesterPush/encoding=DoubleDelta	   20165041 ns/op  2590736 B/op	   12441 allocs/op
BenchmarkIngesterPushErrors/encoding=DoubleDelta   26534724 ns/op  5270153 B/op	   85194 allocs/op
```

After #2000 
```
BenchmarkIngesterPush/encoding=DoubleDelta	   20305149 ns/op  2591191 B/op	   12444 allocs/op
BenchmarkIngesterPushErrors/encoding=DoubleDelta   27959445 ns/op  5276005 B/op	   85189 allocs/op
```

This branch:
```
BenchmarkIngesterPush/encoding=DoubleDelta	   20074468 ns/op  2591331 B/op	   12598 allocs/op
BenchmarkIngesterPushErrors/encoding=DoubleDelta   35415601 ns/op 11072246 B/op	   95055 allocs/op
```