@pracucci 2020-01-21
@bboreham Do you mind rebasing? This stuff has changed in the WAL PR https://github.com/cortexproject/cortex/pull/1103 which has just been merged 😞