@bboreham 2020-01-23
I think the benchmarks show that we don't want to do this so have updated #2004.